+++ tests/docker_cluster.py
@@ -392,17 +392,20 @@ class DockerCluster(BaseCluster):
         return centos_cluster
 
     @staticmethod
-    def _check_for_images(master_image_name, slave_image_name):
+    def _check_for_images(master_image_name, slave_image_name, tag='latest'):
+        master_repotag = '%s:%s' % (master_image_name, tag)
+        slave_repotag = '%s:%s' % (slave_image_name, tag)
         with Client(timeout=180) as client:
             images = client.images()
         has_master_image = False
         has_slave_image = False
         for image in images:
-            if master_image_name in image['RepoTags'][0]:
+            if master_repotag in image['RepoTags']:
                 has_master_image = True
-            if slave_image_name in image['RepoTags'][0]:
+            if slave_repotag in image['RepoTags']:
                 has_slave_image = True
         return has_master_image and has_slave_image
+
 
     def commit_images(self, bare_image_provider, cluster_type):
         self.client.commit(self.master,
